Adds chapter on linear differential forms
This is all previous untracked work, back when I wasn't using git yet,
so this is the initial commit for this branch.

The chapter is complete until the definition of closed forms, but it
still lacks examples and figures. The introduction to the whole chapter
could be also be rephrased in a better and more understandable way.