Adds theorems on closed differential forms
Added counterexample of a closed but not exact form.

Added theorems that prove the exactness of closed forms
* over star domains
* over simply connected sets (Poincaré's lemma)
providing also their definitions, and that of homotopic curves.

Added example on how to compute the potential of an exact form,
using the integral over a segment from a fixed (x_0,y_0) to an
arbitrary (x,y) in the given set.