Adds proper introduction to 1-forms
The introduction to linear differential forms gets completely rewritten:
now it explains where the {dx,dy,dz...} notation arises, from the
differential of scalar functions as elements of the dual space of Rn to
dx,dy... as elements of its basis.

Also added a brief introduction on dual spaces and linear functionals in
general, to familiarize with the concept before introducing the
differentials as linear functionals.