Corrects proof of exactness of a 1-form
Corrected an error in the incremental ratio of F in the limit of the
second to last equation (there was a wrong 'h'), in the theorem which
states that a 1-form is exact iff its integral on a path depends only by
the extreme points of the path.

Also improved here and there some wording.